Change logic in boundary check
The wrapping was off by 1 place so I made the highest number
(i.e., the numbers respectively associated with "Z" and "z") +1
to return the correct encryption (move 3 places down to alphabet) for
letters at the end of the alphabet that needed to be wrapped to be letters
at the beginning of the alphabet like a. A little finicky quick fix but
it gets the job done.